+++ airflow_watchdog/__init__.py
@@ -15,7 +15,7 @@ Install:
     pip install airflow-provider-watchdog
 """
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 
 def get_provider_info() -> dict: